# kv,server,roachpb: avoid error overhead for x-locality comparison