@kvoli 2023-10-30
Benchmarking the comparison of this commit cherry-picked on `relase-23.2` and the predecessor:

```
$ benchdiff --old=90ddb8a960dd11da7c72286268cb42fd656dd3b3  --new=684b9cab0c342ed385c5968c8da4fd5eb46af597 --run=BenchmarkColBatchScan --memprofile --count=20 ./pkg/sql/colflow
checking out '90ddb8a'
building benchmark binaries for '90ddb8a' 1/1 /
checking out '684b9ca'
building benchmark binaries for '684b9ca' 1/1 |

  pkg=1/1 iter=20/20 cockroachdb/cockroach/pkg/sql/colflow \

name                        old time/op    new time/op    delta
ColBatchScan/rows=16-24        123µs ± 6%      76µs ± 5%  -38.27%  (p=0.000 n=19+20)
ColBatchScan/rows=256-24       221µs ± 3%     174µs ± 5%  -21.18%  (p=0.000 n=19+18)
ColBatchScan/rows=4096-24     1.44ms ± 1%    1.41ms ± 1%   -2.66%  (p=0.000 n=19+19)
ColBatchScan/rows=65536-24    22.1ms ± 1%    22.2ms ± 1%     ~     (p=0.201 n=19+19)

name                        old speed      new speed      delta
ColBatchScan/rows=16-24     2.08MB/s ± 6%  3.37MB/s ± 5%  +62.01%  (p=0.000 n=19+20)
ColBatchScan/rows=256-24    18.5MB/s ± 3%  23.5MB/s ± 5%  +26.91%  (p=0.000 n=19+18)
ColBatchScan/rows=4096-24   45.4MB/s ± 1%  46.6MB/s ± 1%   +2.73%  (p=0.000 n=19+19)
ColBatchScan/rows=65536-24  47.4MB/s ± 1%  47.2MB/s ± 1%     ~     (p=0.199 n=19+19)

name                        old alloc/op   new alloc/op   delta
ColBatchScan/rows=16-24       11.3kB ± 1%     7.8kB ± 1%  -31.19%  (p=0.000 n=18+20)
ColBatchScan/rows=256-24      29.5kB ± 0%    27.8kB ± 0%   -5.73%  (p=0.000 n=20+19)
ColBatchScan/rows=65536-24    2.18MB ± 0%    2.20MB ± 1%   +0.86%  (p=0.000 n=19+18)
ColBatchScan/rows=4096-24      192kB ± 0%     204kB ± 7%   +6.32%  (p=0.000 n=18+20)

name                        old allocs/op  new allocs/op  delta
ColBatchScan/rows=16-24          161 ± 1%       108 ± 0%  -32.83%  (p=0.000 n=18+20)
ColBatchScan/rows=256-24         178 ± 0%       157 ± 0%  -11.65%  (p=0.000 n=20+19)
ColBatchScan/rows=4096-24        213 ± 0%       292 ±34%  +36.97%  (p=0.001 n=17+20)
ColBatchScan/rows=65536-24       279 ± 7%       419 ±24%  +50.39%  (p=0.000 n=19+18)

```