@kvoli 2023-10-30
blathers backport 23.2